pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aerich>=0.9.1",
    "asgi-lifespan>=2.1.0",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "beautifulsoup4>=4.13.5",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "lxml>=6.0.1",
    "openai>=1.107.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sqlalchemy[asyncio]>=2.0.43",
    "tortoise-orm>=0.25.1",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "httpx>=0.28.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.12",
]

[tool.ruff]
# 파일 또는 디렉토리 제외
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
]
# 한 줄 최대 길이를 설정합니다. (Black과 일치시킵니다)
line-length = 88

[tool.ruff.lint]
# Ruff가 적용될 코드의 경로를 지정합니다.
select = ["E", "F", "B", "I"]
# E: pycodestyle, F: pyflakes, B: bugbear, I: isort
ignore = []
# 무시할 규칙이 있다면 여기에 추가합니다. 예: ignore = ["E501"] (한 줄 최대 길이)


[tool.aerich]
tortoise_orm = "app.db.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[project.optional-dependencies]
test = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "httpx>=0.27",
  "asgi-lifespan>=2.1.0",
  "tortoise-orm>=0.20.0",
]
